fix(session): isolate Agent worker from ModelsDev runtime
Keep build-time model catalog loading out of the Agent runner static graph so source workers can reach the ready handshake without depending on models.dev.
