+++ .synergy/skill/integrate-llm/SKILL.md
@@ -72,7 +72,7 @@ When changing Agent worker capacity, a higher ceiling must admit queued demand i
 
 Coalesce adjacent streamed tool-input deltas by call ID before IPC, cap each coalesced delta, and preserve cross-call event order. Do not forward one IPC frame per provider token fragment: streamed tool arguments are progress data, while the final AI SDK `tool-call` input remains canonical. Treat one active-turn Bun `heapUsed` heartbeat as a pressure signal rather than proof of retained memory; request a full collection in that worker and require the corresponding post-collection sample to remain over the watermark before terminating the turn.
 
-The executable bootstrap and Agent worker runner static value-import graphs are also product boundaries. Keep `src/index.ts` free of static application imports so compiled worker subcommands do not evaluate the main CLI/server graph. Keep Browser, Tool, Plugin, Plugin Runtime, `packages/browser`, and provider SDK implementations out of the Agent runner graph. Resolve Control Plane preparation dependencies before transfer, and load only the selected built-in provider SDK lazily inside the worker. Update and run `test/session/agent-worker-runtime-boundary.test.ts` whenever the executable entrypoint, runner, LLM, provider, schema, or shared utility import graph changes.
+The executable bootstrap and Agent worker runner static value-import graphs are also product boundaries. Keep `src/index.ts` free of static application imports so compiled worker subcommands do not evaluate the main CLI/server graph. The Agent runner value-import graph may include only worker-safe, side-effect-free schemas and helpers (import `ModelsDev` from `provider/models-schemas.ts`, not `provider/models.ts`). Exclude `provider/models.ts` (runtime/cache/macro dependent), `provider/models-macro.ts` (compile-time external network I/O), Browser, Tool, Plugin, Plugin Runtime, `packages/browser`, and provider SDK implementations from the runner graph. ModelsDev runtime, catalog, and macro remain Control Plane / host-only; no compile-time or module-load external network or other unbounded I/O may execute before the worker `ready`/heartbeat. Resolve Control Plane preparation dependencies before transfer, and load only the selected built-in provider SDK lazily inside the worker. Update and run `test/session/agent-worker-runtime-boundary.test.ts` when this boundary changes.
 
 Provider SSE protection is a per-event parser bound, not a total response, transport chunk, or process-memory limit. Keep code identifiers, error names, tests, and architecture documentation explicit about `SSE event parser bound` semantics. Test LF and CRLF event delimiters across chunk boundaries, including consecutive events exactly at the bound.
