+++ docs/architecture/llm-loop.md
@@ -281,7 +281,7 @@ Tool definitions cross the Agent worker boundary as plain JSON Schema data. AI S
 
 The external `AgentTurn` boundary transfers immutable turn snapshots through a versioned, schema-validated protocol. Before transfer, the Control Plane runs prompt and parameter plugin hooks, resolves the built-in provider profile, strips executable provider options, freezes provider request timeouts into the worker plan, and retains optional Context Usage provenance locally. Requests are capped at 64 MiB and paged through acknowledged 1 MiB chunks rather than one unbounded IPC object. Worker event frames are Synergy-owned projections rather than raw AI SDK stream objects: lifecycle events retain only fields consumed by the Control Plane, so provider request bodies, response diagnostics, warnings, Context Usage provenance, and Context Usage drafts do not cross the process boundary. Frames are capped at 2 MiB, text/reasoning deltas coalesce up to 16 ms or 32 KiB, and adjacent tool-input deltas coalesce by call ID up to 32 KiB while preserving cross-call event order. Each emitted frame remains the only buffered frame until the Control Plane consumer acknowledges it.
 
-The executable entrypoint is a thin bootstrap with no static application dependencies. It selects the main CLI/server, Agent runner, Policy runner, or Plugin runner through one mutually exclusive dynamic import. The Agent runner then has its own enforced static runtime boundary: its value-import graph may include the LLM loop, serializable schemas, provider reconstruction, and the selected provider SDK, but it may not include Browser, Tool, Plugin, Plugin Runtime, or `packages/browser` implementations. Provider SDKs load lazily after the worker receives a model plan; unselected SDKs do not enter a source-mode worker process. Browser names and JSON Schemas can cross as `ToolCatalog.Definition[]`, while Browser sessions, Playwright, tool callbacks, plugin discovery, Host Services, configuration writers, and permission state remain in the Control Plane or their owning tool runtime.
+The executable entrypoint is a thin bootstrap with no static application dependencies. It selects the main CLI/server, Agent runner, Policy runner, or Plugin runner through one mutually exclusive dynamic import. The Agent runner then has its own enforced static runtime boundary: its value-import graph may include only worker-safe, side-effect-free schemas and helpers (`provider/models-schemas.ts`), the LLM loop, serializable schemas, provider reconstruction, and lazily loaded selected-provider SDK. It must not include `provider/models.ts` (runtime/cache/macro dependent), `provider/models-macro.ts` (compile-time external network I/O), Browser, Tool, Plugin, Plugin Runtime, or `packages/browser` implementations. ModelsDev runtime, catalog, and macro remain Control Plane / host-only; no compile-time or module-load external network or other unbounded I/O may execute before the worker `ready`/heartbeat. When this import boundary changes, update and run `test/session/agent-worker-runtime-boundary.test.ts`.
 
 The pool is elastic. `agentWorkers` is its parallelism ceiling; demand starts workers up to that ceiling instead of preallocating the full pool. `agentWorkerMinIdle` optionally maintains a warm reserve, and workers above the reserve retire after `agentWorkerIdleTimeoutMs`. Raising the ceiling admits queued demand without eagerly filling unused capacity. Lowering it releases excess idle workers immediately and marks excess active workers to retire only after their current turns reach the worker-release boundary. Queue count and aggregate bytes are bounded independently. Cancellation, heartbeat timeout, startup circuit breaking, crash containment, soft RSS/heap recycling, hard RSS/heap containment, turn-count recycling, and parent-death cleanup affect only the owned turn.
