+++ packages/synergy/src/provider/provider.ts
@@ -5,7 +5,7 @@ import { mapValues, mergeDeep, omit, pickBy, sortBy } from "remeda"
 import { NoSuchModelError, type Provider as SDK } from "ai"
 import { Log } from "../util/log"
 import { BunProc } from "../util/bun"
-import { ModelsDev } from "./models"
+import { ModelsDev } from "./models-schemas"
 import { NamedError } from "@ericsanchezok/synergy-util/error"
 import { Auth } from "./api-key"
 import { Env } from "../util/env"
